requirements.txt
dnspython==2.2.1
geoip2==4.6.0
scapy==2.4.5
colorama==0.4.6
requests==2.28.2
